[Fluid] Implement JIT parser modernisation plan phase 3 (#912)
* [Fluid] Add register allocator RAII interface; Fixed mis-positioned register deallocation
* Add parser expression wrapper scaffolding (#907)
* Implement control flow graph for parser jumps (#908)
* [Fluid]  LuaJIT Parser Stage 4 C++ Enhancements (#910)

Phase 1.1 - ExpressionValue helper methods:
- Added to_val() for partial discharge (like expr_toval)
- Added discharge_nobranch() for discharge without jumps
- Added store_to() for storing values (like bcemit_store)
- Added discharge_to_any_reg() combining discharge + to_any_reg

Phase 1.2 - IrEmissionContext class:
- Created context class bundling RegisterAllocator, ControlFlowGraph, and FuncState
- Provides unified API: allocator(), cfg(), state()
- Simplifies function signatures for gradual migration

Phase 2.2 - Replace manual table duplication:
- Removed duplicate_index_base() helper function
- Updated emit_compound_assignment() to use RegisterAllocator::duplicate_table_operands()
- Updated emit_if_empty_assignment() to use RegisterAllocator::duplicate_table_operands()
- Updated emit_update_expr() to use RegisterAllocator::duplicate_table_operands()

Migrated ir_emitter.cpp functions to use ExpressionValue APIs:
- emit_if_empty_assignment(): expr_discharge+expr_toanyreg → discharge_to_any_reg()
- emit_update_expr(): expr_discharge+expr_toanyreg → discharge_to_any_reg(), bcreg_reserve → allocator.reserve()
- emit_ternary_expr(): expr_discharge+expr_toanyreg → discharge_to_any_reg(), 2x expr_discharge → discharge()

**Agent 1 - ir_emitter.cpp (16 functions, 21 calls):**
- Migrated emit_return_stmt, emit_local_function_stmt, emit_numeric_for_stmt
- Migrated emit_generic_for_stmt, emit_defer_stmt, emit_vararg_expr
- Migrated emit_member_expr, emit_index_expr, emit_call_expr
- Migrated emit_table_expr, emit_function_expr
- Migrated emit_function_lvalue, emit_lvalue_expr
- Migrated materialise_to_next_reg, materialise_to_reg
- Fixed compilation errors from incorrect method names

**Agent 2 - parse_operators.cpp (9 functions, 47 calls):**
- Migrated bcemit_arith (4 calls), bcemit_comp (6 calls)
- Migrated bcemit_binop_left (7 calls), bcemit_shift_call_at_base (5 calls)
- Migrated bcemit_bit_call (3 calls), bcemit_unary_bit_call (6 calls)
- Migrated bcemit_presence_check (3 calls), bcemit_binop (10 calls)
- Migrated bcemit_unop (3 calls)

**Migration Patterns Applied:**
- expr_discharge + expr_toanyreg → discharge_to_any_reg()
- expr_toanyreg → discharge_to_any_reg()
- expr_toreg → to_reg()
- expr_tonextreg → to_next_reg()
- bcreg_reserve → allocator.reserve()

**Agent 3 - parse_expr.cpp (14 functions, 33 calls):**
- Migrated expr_index, expr_field, expr_bracket
- Migrated expr_table (2 calls), parse_params
- Migrated expr_list, parse_args, expr_primary_with_context (4 calls)
- Migrated expr_simple_with_context, inc_dec_op (6 calls)
- Migrated expr_shift_chain (3 calls), expr_binop (10 calls)
- Migrated expr_next

**Agent 4 - parse_stmt.cpp (10 functions, 29 calls):**
- Migrated assign_hazard, assign_adjust (3 calls)
- Migrated assign_if_empty (10 calls), assign_compound (5 calls)
- Migrated parse_local (2 calls), snapshot_return_regs (2 calls)
- Migrated parse_defer (3 calls), parse_return (2 calls)
- Migrated parse_for_num (2 calls), parse_for_iter

**Migration Patterns Applied:**
- expr_discharge + expr_toanyreg → discharge_to_any_reg()
- expr_toanyreg → discharge_to_any_reg()
- expr_tonextreg → to_next_reg()
- expr_toreg → to_reg()
- expr_toval → to_val()
- bcreg_reserve → allocator.reserve()
- bcreg_free → allocator.release_register()

Internalized flag bit manipulation into ExpressionValue methods:
- has_flag(): Direct bit test (flags & Flag) != ExprFlag::None
- set_flag(): Direct bit set (flags |= Flag)
- clear_flag(): Direct bit clear (flags &= ~Flag)
- consume_flag(): Test, clear if set, return result

Implemented comprehensive debug verification and lightweight tracing hooks
for the register allocator and control-flow graph, completing Phase 3.

RegisterAllocator enhancements:
- Added constructor to track initial freereg state for debug builds
- Implemented verify_balance() to check register depth consistency
- Implemented verify_no_leaks() to detect unreleased temporary registers
- Added trace_allocation() and trace_release() for register ops tracing
- Tracing controlled by LJ_TRACE_REGALLOC compile-time flag

ControlFlowGraph enhancements:
- Added trace_edge_creation() to log edge construction
- Added trace_edge_patch() to log jump patching operations
- Added trace_edge_append() to log edge list modifications
- Integrated tracing into make_edge(), patch_edge(), and append_edge()
- Tracing controlled by LJ_TRACE_CFG compile-time flag

IrEmitter verification points:
- Added verify_no_leaks() call at function exit (emit_chunk)
- Added verify_no_leaks() call at while loop exit
- Added verify_no_leaks() call at repeat loop exit
- Leverages existing ensure_register_balance() at statement boundaries

All debug features guarded by #if LJ_DEBUG and compiled successfully
in Debug build. Tracing can be enabled via -DLJ_TRACE_REGALLOC and
-DLJ_TRACE_CFG flags for detailed debugging output.

* Remove unused and flawed verify_balance() method

The verify_balance() method was declared but never called, and its design
was fundamentally flawed for the verification points in this implementation.
Creating a new RegisterAllocator instance captures the current freereg as
initial state, which would make verify_balance() always pass if called
immediately.

The verify_no_leaks() method already provides the necessary verification
by checking that freereg doesn't exceed nactvar (i.e., no temporary
registers leaked), which covers the balance checking requirement.

Removed:
- verify_balance() method declaration and implementation
- initial_freereg member variable (no longer needed)
- Constructor initialization of initial_freereg

This simplifies the debug verification interface to only the methods
that are actually used and correctly designed.

* Integrate RegisterAllocator tracing into allocation/release paths

The trace_allocation() and trace_release() methods were declared and
implemented but never called. This integrates them into the actual
code paths to provide functional tracing when LJ_TRACE_REGALLOC is
defined.

Changes:
- Added trace_allocation() call in reserve_slots() after allocation
- Added trace_release() call in release_span_internal() after release
- Tracing now mirrors the ControlFlowGraph pattern where trace methods
  are called from the actual operations

This matches the ControlFlowGraph tracing integration where trace
methods are properly called from make_edge(), append_edge(), and
patch_edge() operations.

When compiled with -DLJ_TRACE_REGALLOC, the RegisterAllocator will now
output register allocation and release events to stderr for debugging.

+++ src/fluid/luajit-2.1/src/lj_record.cpp
@@ -2245,7 +2245,9 @@ void lj_record_ins(jit_State* J)
    switch (bcmode_c(op)) {
    case BCMvar:
       copyTV(J->L, rcv, &lbase[rc]); ix.key = rc = getslot(J, rc); break;
-   case BCMpri: setpriV(rcv, ~rc); ix.key = rc = TREF_PRI(IRT_NIL + rc); break;
+   case BCMpri: 
+      setpriV(rcv, ~uint64_t(rc)); 
+      ix.key = rc = TREF_PRI(IRT_NIL + rc); break;
    case BCMnum: {
       cTValue* tv = proto_knumtv(J->pt, rc);
       copyTV(J->L, rcv, tv); ix.key = rc = tvisint(tv) ? lj_ir_kint(J, intV(tv)) :
